Promote queue mutation DTOs to HurrahTv.Shared
The five queue-endpoint request records (QueueStatusUpdate,
PositionUpdate, SentimentUpdate, ProgressUpdate, SeenRequest) lived as
private records at the bottom of QueueEndpoints.cs, while ApiClient sent
anonymous objects whose property names just happened to match. A field
rename or typo on either side would fail silently at runtime.

Move the records into HurrahTv.Shared/Models/QueueModels.cs so both
sides reference one definition, and swap the five anonymous payloads in
ApiClient (UpdateStatusAsync, UpdateSentimentAsync,
UpdateQueuePositionAsync, UpdateProgressAsync, PostStatusAsync) for the
typed records. Positional records serialize on their parameter names by
default in System.Text.Json, so the wire shape is unchanged.

QueueEndpoints.cs now declares only its routes. The remaining
anonymous-object payloads in ApiClient belong to sentiment / admin /
auth endpoint families and are out of scope here.

Closes #64.
