+++ HurrahTv.Api/Endpoints/QueueEndpoints.cs
@@ -201,10 +201,4 @@ public static class QueueEndpoints
             return true; // malformed payload — don't hide
         }
     }
-
-    public record QueueStatusUpdate(QueueStatus Status);
-    public record PositionUpdate(int Position);
-    public record SentimentUpdate(int? Sentiment);
-    public record ProgressUpdate(int? Season, int? Episode);
-    public record SeenRequest(int TmdbId, string MediaType, string Title, string PosterPath, string AvailableOnJson, string BackdropPath = "");
 }
